:recycle: refactor: Move processZipFile function to FileOperations class
