+++ tofhir-server/src/main/scala/io/tofhir/server/endpoint/SchemaDefinitionEndpoint.scala
@@ -1,6 +1,6 @@
 package io.tofhir.server.endpoint
 
-import akka.http.scaladsl.model.{ContentTypes, StatusCodes}
+import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import com.typesafe.scalalogging.LazyLogging
@@ -10,26 +10,18 @@ import io.tofhir.server.endpoint.SchemaDefinitionEndpoint.{SEGMENT_IMPORT, SEGME
 import io.tofhir.common.model.Json4sSupport._
 import io.tofhir.server.model.{ImportSchemaSettings, InferTask}
 import io.tofhir.engine.util.FhirMappingJobFormatter.formats
-import io.tofhir.server.common.model.{BadRequest,InternalError, ResourceNotFound, ToFhirRestCall}
+import io.tofhir.server.common.model.{BadRequest, InternalError, ResourceNotFound, ToFhirRestCall}
 import io.tofhir.server.endpoint.MappingContextEndpoint.ATTACHMENT
 import io.onfhir.api.Resource
 import io.tofhir.server.repository.mapping.IMappingRepository
 import io.tofhir.server.repository.schema.ISchemaRepository
 import io.tofhir.server.service.SchemaDefinitionService
-import akka.http.scaladsl.model.{HttpEntity, HttpResponse}
 import akka.http.scaladsl.server.directives.FileInfo
-import io.onfhir.exception.InitializationException
-import io.onfhir.util.OnFhirZipInputStream
-import org.apache.commons.io.input.BOMInputStream
-import org.json4s.JsonAST.JObject
-import org.json4s.jackson.JsonMethods
+import io.tofhir.server.util.FileOperations
 
-import java.io.{File, InputStreamReader, Reader}
-import java.nio.file.{Files, Path}
-import java.util.zip.ZipEntry
-import scala.concurrent.Future
+import java.io.File
+import java.nio.file.Files
 import scala.util.{Failure, Success}
-import scala.collection.mutable
 
 class SchemaDefinitionEndpoint(schemaRepository: ISchemaRepository, mappingRepository: IMappingRepository) extends LazyLogging {
 
@@ -218,7 +210,7 @@ class SchemaDefinitionEndpoint(schemaRepository: ISchemaRepository, mappingRepos
     post {
       storeUploadedFile("file", createTempFile) {
         case (_, file) =>
-          val zipProcessingResult = processZipFile(file.toPath)
+          val zipProcessingResult = FileOperations.processZipFile(file.toPath)
           onComplete(zipProcessingResult) {
             case Success(result) =>
               complete(
@@ -244,55 +236,6 @@ class SchemaDefinitionEndpoint(schemaRepository: ISchemaRepository, mappingRepos
       val tempFile = Files.createTempFile(fileInfo.fileName, ".tmp").toFile
       tempFile.deleteOnExit()
       tempFile
-  }
-
-  /**
-   * Processes a ZIP file containing FHIR structure definitions and extracts the resources.
-   *
-   * This method reads the ZIP file, parses each JSON resource, and collects them into a sequence of
-   * `Resource` objects.
-   *
-   * @param zipFilePath The path to the ZIP file to be processed.
-   * @return A Future containing a sequence of `Resource` objects extracted from the ZIP file.
-   */
-  private def processZipFile(zipFilePath: Path): Future[Seq[Resource]] = Future {
-    /**
-     * Parses a JSON resource from the provided reader.
-     *
-     * This method attempts to parse a JSON resource from the given `Reader`. If the path does not end with ".json",
-     * or if the JSON parsing fails, appropriate exceptions are thrown.
-     *
-     * @param reader The reader to read the JSON data from.
-     * @param path The file path from which the resource is being read.
-     * @return The parsed `Resource` object from the JSON data.
-     *
-     * @throws InternalError If there is a problem parsing the JSON data from the path.
-     * @throws BadRequest If the file path does not end with ".json".
-     */
-    def parseResource(reader: Reader, path: String): Resource = {
-      if (path.endsWith(".json"))
-        try {
-          JsonMethods.parse(reader).asInstanceOf[JObject]
-        }
-        catch {
-          case e: Exception =>
-            throw InternalError("JSON parsing problem",s"Cannot parse resource from path $path!", Some(e))
-        }
-      else
-        throw BadRequest("Invalid JSON!",s"Cannot read resource from path $path, it should be JSON file!")
-    }
-
-    val zipStream = new OnFhirZipInputStream(Files.newInputStream(zipFilePath))
-    val resources: mutable.ListBuffer[Resource] = new mutable.ListBuffer[Resource]
-    var zipEntry: ZipEntry = zipStream.getNextEntry
-
-    while (zipEntry != null) {
-      val reader = new InputStreamReader(BOMInputStream.builder.setInputStream(zipStream).get(), "UTF-8")
-      resources.append(parseResource(reader, zipEntry.getName))
-      zipStream.closeEntry()
-      zipEntry = zipStream.getNextEntry
-    }
-    resources.toSeq
   }
 }
